Split options and configuration extensions into separate files.
