+++ src/OpenTelemetry/Metrics/PeriodicExportingMetricReaderOptions.cs
@@ -2,7 +2,6 @@
 // SPDX-License-Identifier: Apache-2.0
 
 using Microsoft.Extensions.Configuration;
-using OpenTelemetry.Internal;
 
 namespace OpenTelemetry.Metrics;
