+++ src/OpenTelemetry/Trace/BatchExportActivityProcessorOptions.cs
@@ -3,7 +3,6 @@
 
 using System.Diagnostics;
 using Microsoft.Extensions.Configuration;
-using OpenTelemetry.Internal;
 
 namespace OpenTelemetry.Trace;
